[REFACTOR] 서비스 DTO 분리
