+++ src/main/java/com/soptie/server/routine/service/happiness/HappinessSubRoutineService.java
@@ -1,4 +1,4 @@
-package com.soptie.server.routine.service;
+package com.soptie.server.routine.service.happiness;
 
 import com.soptie.server.routine.entity.happiness.HappinessRoutine;
 import com.soptie.server.routine.entity.happiness.HappinessSubRoutine;
